app/requirements.txt
gradio==4.44.0
openai-whisper==20240930
xgboost==2.1.3
scikit-learn==1.5.2
joblib>=1.3.0
torch>=2.0.0
numpy>=1.24.0,<2.0
librosa>=0.10.0
matplotlib>=3.7.0
ffmpeg-python>=0.2.0
imageio-ffmpeg>=0.4.9
